validatioon added

+++ app/Http/Controllers/country/CountryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\country;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Model\CountryModel;
+use Validator;
 
 class CountryController extends Controller
 {
@@ -15,23 +16,46 @@ class CountryController extends Controller
 
     public function countryById($id)
     {
-        return \response()->json(CountryModel::find($id),200);
+        $country= CountryModel::find($id);
+        if(is_null($country))
+        {
+            return \response()->json(["message"=>"Record not found!"],404);
+        }
+        return \response()->json($country,200);
     }
 
     public function countrySave(Request $request)
     {
+        $validation=[
+            'name'=>'required|min:3',
+            'iso'=>'required|min:2|max:2',
+        ];
+        $validator = Validator::make($request->all(), $validation);
+        if($validator->fails()){
+            return \response()->json($validator->errors(),400);
+        }
         $country = CountryModel::create($request->all());
         return \response()->json($country,201);
     }
 
-    public function countryUpdate(Request $request, CountryModel $countryid)
+    public function countryUpdate(Request $request, $id)
     {
+        $country= CountryModel::find($id);
+        if(is_null($country))
+        {
+            return \response()->json(["message"=>"Record not found!"],404);
+        }
         $countryid->update($request->all());
         return \response()->json($countryid,200);
     }
 
-    public function countryDelete(Request $request, CountryModel $countryid)
+    public function countryDelete(Request $request, $id)
     {
+        $country= CountryModel::find($id);
+        if(is_null($country))
+        {
+            return \response()->json(["message"=>"Record not found!"],404);
+        }
         $countryid->delete();
         return \response()->json(null,204);
     }
